+++ src/igor/igor/ui/dialogs/iDialog.h
@@ -178,8 +178,9 @@ namespace igor
 
         \param dialogCloseDelegate the delegate to call after the dialog was closed
         \param modal open this dialog as modal
+        \param reset if true reset dialog and all it's children (usefull when reusing an existing dialog)
         */
-        virtual void open(iDialogCloseDelegate dialogCloseDelegate = iDialogCloseDelegate(), bool modal = false);
+        virtual void open(iDialogCloseDelegate dialogCloseDelegate = iDialogCloseDelegate(), bool modal = false, bool reset = true);
 
         /*! closes the dialog
          */