+++ src/igor/igor/ui/dialogs/iDialogMenu.h
@@ -109,8 +109,9 @@ namespace igor
 
         \param dialogCloseDelegate the delegate to call after the dialog was closed
         \param modal if true open this dialog as modal
+        \param reset if true reset dialog and all it's children
         */
-        virtual void open(iDialogCloseDelegate dialogCloseDelegate = iDialogCloseDelegate(), bool modal = false) override;
+        virtual void open(iDialogCloseDelegate dialogCloseDelegate = iDialogCloseDelegate(), bool modal = false, bool reset = true) override;
 
         /*! clears the widget back to default
         */